+++ backend/src/tipovehiculo/tipovehiculo.controler.ts
@@ -23,7 +23,7 @@ async function findAll(req: Request, res: Response) {
     const tiposVehiculos = await em.find(TipoVehiculo, {})
     res.status(200).json({ message: 'Tipos de Vehiculos encontrados', data: tiposVehiculos })
   } catch (error: any) {
-    res.status(500).json({ message: error.message })
+    res.status(500).json({ message: 'No se encontraron tipos de vehiculos', data: error })
   }
 }
 
@@ -33,7 +33,7 @@ async function findOne(req: Request, res: Response) {
     const tipoVehiculo = await em.findOneOrFail(TipoVehiculo, { id })
     res.status(200).json({ message: 'Tipo de vehiculo encontrado', data: tipoVehiculo })
   } catch (error: any) {
-    res.status(500).json({ message: error.message })
+    res.status(500).json({ message: 'No se encontro el tipo de vehiculo', data: error })
   }
 }
 
@@ -44,31 +44,39 @@ async function add(req: Request, res: Response) {
     await em.flush()
     res.status(201).json({ message: 'Se cargo nuevo tipo de vehiculo', data: tipoVehiculoNuevo })
   } catch (error: any) {
-    res.status(500).json({ message: error.message })
+    res.status(500).json({ message: 'No se pudo cargar el nuevo tipo de vehiculo', data: error })
   }
 }
 
 async function update(req: Request, res: Response) {
   try {
     const id = Number.parseInt(req.params.id)
+    const tipoVehiculoExistente = await em.findOne(TipoVehiculo, { id })
+    if (!tipoVehiculoExistente) {
+      return res.status(404).json({ message: 'El tipo de vehiculo no existe' })
+    }
     req.body.sanitizedInput.id=req.params.id
     const tipoVehiculoModificado = em.getReference(TipoVehiculo, id)
     em.assign(tipoVehiculoModificado, req.body.sanitizedInput)
     await em.flush()
     res.status(200).json({ message: 'Tipo de Vehiculo actualizado correctamente' })
   } catch (error: any) {
-    res.status(500).json({ message: error.message })
+    res.status(500).json({ message: 'No se pudo actualizar el tipo de vehiculo', data: error })
   }
 }
 
 async function remove(req: Request, res: Response) {
   try {
     const id = Number.parseInt(req.params.id)
+    const tipoVehiculoExistente = await em.findOne(TipoVehiculo, { id })
+    if (!tipoVehiculoExistente) {
+      return res.status(404).json({ message: 'El tipo de vehiculo no existe' })
+    }
     const tipoVehiculoBorrar = em.getReference(TipoVehiculo, id)
     await em.removeAndFlush(tipoVehiculoBorrar)
     res.status(200).send({ message: 'Tipo de vehiculo eliminado correctamente' })
   } catch (error: any) {
-    res.status(500).json({ message: error.message })
+    res.status(500).json({ message: 'No se pudo eliminar el tipo de vehiculo', data: error })
   }
 }
